git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@552 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ savant.data/src/savant/data/DataTableModel.java
@@ -232,7 +232,7 @@ public class DataTableModel extends AbstractTableModel {
                          ((ContinuousRecord) datum).setPosition(Integer.parseInt((String) value));
                          break;
                      case 1:
-                         ((ContinuousRecord) datum).getValue().setValue(Double.parseDouble((String) value));
+                         ((ContinuousRecord) datum).getValue().setValue(Float.parseFloat((String) value));
                          break;
                  }
              case INTERVAL_GENERIC:
